Track vao binding
