+++ lib/context.js
@@ -2,7 +2,7 @@ import { generateId, log, error } from './utils';
 import { constants } from './constants';
 import { Program } from './program';
 import { ArrayBuffer, ElementArrayBuffer } from './buffer';
-import { VertexArrayObject } from './vao';
+import { VertexArrayObject } from './vertex-array-object';
 import { BLACK } from './color';
 
 const {
@@ -105,7 +105,7 @@ export class Context {
         return new ElementArrayBuffer(this);
     }
 
-    createVAO() {
+    createVertexArrayObject() {
         return new VertexArrayObject(this);
     }
 }
